added Anim class to optimizer, many opt fixes!
Added a Anim class which uses awt to animate the optimization process.
It follows the iterations and paints each one after the other (with
both the rectangles and the already selected points).

In the process refactored the optimizer code to use a Iterator, so
that it can be easily triggered from outside the class (as required
by the Anim class).

Also made a set of fixes in both the branch and permutations methods.
Lots of possible permutations were missing and the result was never
accurate.

Added lots of tests for the optimizer with basic files (3, 5 and 10
points), and others for branch and permutations.