added soaringweb airspace data parsing.
This commit adds the ability to parse for soaringweb releases. The only
available functionality for now is finding the openair release files,
which can then be parsed to retrieved the actual airspace data.

As part of a future commit we'll refactor this and the other packages to
perform get/put operations consistently across airspace, airfield and
waypoint data.

Implements #19.