add cli pkg, add airspace-get, added soaringweb.
adds a new cli package holding all the cli command implementations, for
now only airspace-get.

added an Airspacer interface with get/put methods, and an implementation
of the Get method for soaringweb.

adds a new plugin interface, to be implemented by all plugins.

added a mock package for dummy implementations of all interfaces (useful
for testing).