basic entry page, airfield/waypoint gmap display.
adds a simple index.html page for entry in the site, with an initial
google maps based display for airfields and waypoints.

it's currently based on managing markers manually. later we should check
if managing directly using the google maps data layer (which handles
geojson nicely) is possible. tried but couldn't figure out how to get
additional maps for a single Point feature (for the labels).

another possibility is to rely on webgl directly.

other things:
- inverted lat/lon in geojson.

Fixes #46.