adds flight optimizer/scorer and montecarlo.
adds the flight optimizer and scorer interfaces.

the first returns the score and distance for a given flight, allowing
multiple implementations of the optimization algorithm (montecarlo,
simulated annealing, genetic, branch/bound, ...). actual implementation
based on montecarlo is included.

the second is the actual scoring interface, which takes a set of turn
points and returns the score points for the competition (netcoupe,
online contest, ...). netcoupe scoring implementation is included.

Fixes #48, #85.